Rename transactionResult to txBroadcaster
