+++ src/modules/token/fungible/components/SendStatus/status.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { isEmpty } from 'src/utils/helpers';
 import { txStatusTypes } from '@transaction/configuration/txStatus';
 import { PrimaryButton } from 'src/theme/buttons';
-import TransactionResult from '@transaction/manager/transactionResult';
+import TransactionResult from '@transaction/manager/txBroadcaster';
 import {
   getTransactionStatus,
   statusMessages,
