+++ src/modules/wallet/components/registerMultisigStatus/status.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import TransactionResult from '@transaction/manager/transactionResult';
+import TransactionResult from '@transaction/manager/txBroadcaster';
 import { statusMessages, getTransactionStatus } from '@transaction/configuration/statusConfig';
 import ProgressBar from '../RegisterMultisigView/ProgressBar';
 import styles from './styles.css';
